Web Service
In this project you will create a web service that checks the strength
of a password.

The endpoint of your web service will be
cs34.cs.sjsu.edu/YOUR_USERNAME/public_html/p8_password.php?pw=XXXXXXX

Where the XXXXXX represents a password that I want to check.

Your program will return 3 pieces of data formatted in a json array.

1. Is the password in the list of the 10,000 most common passwords?
(true or false)

2. If so, then where in the list is it ranked? (a number or -1 if it is
not on the list)

3. A number from 0 to 5 representing the strength of the password. 0 is
very weak, while 5 is strong.

To calculate the strength, give a password one point for each of the
following criteria:

1. 8 or more characters
2. contains uppercase letters
3. contains lowercase letters
4. contains numbers
5. contains other characters